# clients/go/api/openapi.yaml
openapi: 3.0.0
info:
  description: "API REST gamificada para gerenciamento de tarefas com IA, XP, levels\
    \ e sistema de recompensas"
  title: Gamified Task Management API
  version: 1.0.0
servers:
- description: Servidor de Desenvolvimento
  url: http://localhost:3000
security:
- bearerAuth: []
...
paths:
...
  /api/leaderboard:
    get:
      operationId: getLeaderboard
      parameters:
      - description: Número de usuários no ranking
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 50
          minimum: 1
          type: integer
        style: form
      - description: Período do ranking
        explode: true
        in: query
        name: period
        required: false
        schema:
          default: all
          enum:
          - all
          - weekly
          - monthly
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getLeaderboard_200_response'
          description: Leaderboard dos top usuários
      summary: Obter leaderboard público
      tags:
      - Gamification
...
  /api/gamification/preferences:
    put:
      operationId: updateGamificationPreferences
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateGamificationPreferences_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/updateGamificationPreferences_200_response'
          description: Preferências atualizadas
      security:
      - bearerAuth: []
      summary: Atualizar preferências de gamificação
      tags:
      - Gamification
...
  /api/webhooks/share:
    post:
      operationId: processShareWebhook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/processShareWebhook_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/processShareWebhook_200_response'
          description: Webhook processado
      summary: Webhook para processar compartilhamentos
      tags:
      - Webhooks
...
  /api/users/stats:
    get:
      operationId: getUserStats
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getUserStats_200_response'
          description: Estatísticas do sistema
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado - privilégios de admin necessários
      summary: Obter estatísticas gerais do sistema (Admin only)
      tags:
      - Users
...
components:
  schemas:
...
    Error:
      example:
        success: false
        message: Erro de validação
      properties:
        success:
          example: false
          type: boolean
        message:
          example: Erro de validação
          type: string
      type: object
...
    getLeaderboard_200_response_data_leaderboard_inner:
      example:
        badges: 8
        totalXP: 1450
        currentStreak: 12
        tasksCompleted: 87
        level: 15
        rank: 1
        username: GamerPro123
      properties:
        rank:
          example: 1
          type: number
        username:
          example: GamerPro123
          type: string
        level:
          example: 15
          type: number
        totalXP:
          example: 1450
          type: number
        tasksCompleted:
          example: 87
          type: number
        currentStreak:
          example: 12
          type: number
        badges:
          example: 8
          type: number
      type: object
    getLeaderboard_200_response_data:
      example:
        leaderboard:
        - badges: 8
          totalXP: 1450
          currentStreak: 12
          tasksCompleted: 87
          level: 15
          rank: 1
          username: GamerPro123
        - badges: 8
          totalXP: 1450
          currentStreak: 12
          tasksCompleted: 87
          level: 15
          rank: 1
          username: GamerPro123
        period: all
        total: 10
      properties:
        leaderboard:
          items:
            $ref: '#/components/schemas/getLeaderboard_200_response_data_leaderboard_inner'
          type: array
        period:
          example: all
          type: string
        total:
          example: 10
          type: number
      type: object
    getLeaderboard_200_response:
      example:
        data:
          leaderboard:
          - badges: 8
            totalXP: 1450
            currentStreak: 12
            tasksCompleted: 87
            level: 15
            rank: 1
            username: GamerPro123
          - badges: 8
            totalXP: 1450
            currentStreak: 12
            tasksCompleted: 87
            level: 15
            rank: 1
            username: GamerPro123
          period: all
          total: 10
        success: true
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/getLeaderboard_200_response_data'
      type: object
...
    updateGamificationPreferences_request:
      properties:
        class:
          enum:
          - Aventureiro
          - Guerreiro
          - Mago
          - Ladino
          - Paladino
          - Explorador
          example: Guerreiro
          type: string
        theme:
          enum:
          - medieval
          - scifi
          - modern
          - fantasy
          example: medieval
          type: string
        autoShare:
          example: false
          type: boolean
        shareOnLevelUp:
          example: true
          type: boolean
        shareOnBadge:
          example: true
          type: boolean
      type: object
    updateGamificationPreferences_200_response_data:
      example:
        preferences: "{}"
      properties:
        preferences:
          type: object
      type: object
    updateGamificationPreferences_200_response:
      example:
        data:
          preferences: "{}"
        success: true
        message: message
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/updateGamificationPreferences_200_response_data'
      type: object
...
    processShareWebhook_request:
      properties:
        event_type:
          type: string
        user_id:
          type: string
        data:
          type: object
      type: object
    processShareWebhook_200_response:
      example:
        success: true
        message: message
        timestamp: timestamp
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
      type: object
...
    getUserStats_200_response_data_users:
      example:
        total: 150
        inactive: 10
        active: 140
        admins: 5
      properties:
        total:
          example: 150
          type: integer
        active:
          example: 140
          type: integer
        inactive:
          example: 10
          type: integer
        admins:
          example: 5
          type: integer
      type: object
    getUserStats_200_response_data_tasks:
      example:
        total: 1250
        inProgress: 150
        overdue: 45
        pending: 300
        completed: 800
      properties:
        total:
          example: 1250
          type: integer
        completed:
          example: 800
          type: integer
        pending:
          example: 300
          type: integer
        inProgress:
          example: 150
          type: integer
        overdue:
          example: 45
          type: integer
      type: object
    getUserStats_200_response_data_mostActiveUsers_inner:
      example:
        taskCount: 0
        name: name
        userId: userId
        email: email
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
        taskCount:
          type: integer
      type: object
    getUserStats_200_response_data:
      example:
        mostActiveUsers:
        - taskCount: 0
          name: name
          userId: userId
          email: email
        - taskCount: 0
          name: name
          userId: userId
          email: email
        users:
          total: 150
          inactive: 10
          active: 140
          admins: 5
        tasks:
          total: 1250
          inProgress: 150
          overdue: 45
          pending: 300
          completed: 800
      properties:
        users:
          $ref: '#/components/schemas/getUserStats_200_response_data_users'
        tasks:
          $ref: '#/components/schemas/getUserStats_200_response_data_tasks'
        mostActiveUsers:
          items:
            $ref: '#/components/schemas/getUserStats_200_response_data_mostActiveUsers_inner'
          type: array
      type: object
    getUserStats_200_response:
      example:
        data:
          mostActiveUsers:
          - taskCount: 0
            name: name
            userId: userId
            email: email
          - taskCount: 0
            name: name
            userId: userId
            email: email
          users:
            total: 150
            inactive: 10
            active: 140
            admins: 5
          tasks:
            total: 1250
            inProgress: 150
            overdue: 45
            pending: 300
            completed: 800
        success: true
      properties:
        success:
          example: true
          type: boolean
        data:
          $ref: '#/components/schemas/getUserStats_200_response_data'
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http